--- kaldb/src/main/java/com/slack/kaldb/histogram/NoOpHistogramImpl.java
@@ -1,34 +1,0 @@
-package com.slack.kaldb.histogram;
-
-import java.util.Collections;
-import java.util.List;
-
-public class NoOpHistogramImpl implements Histogram {
-  private int count;
-
-  public NoOpHistogramImpl() {
-    count = 0;
-  }
-
-  @Override
-  public void add(long value) {
-    count++;
-  }
-
-  @Override
-  public void mergeHistogram(List<HistogramBucket> mergeBuckets) {
-    for (HistogramBucket bucket : mergeBuckets) {
-      count += bucket.getCount();
-    }
-  }
-
-  @Override
-  public List<HistogramBucket> getBuckets() {
-    return Collections.emptyList();
-  }
-
-  @Override
-  public long count() {
-    return count;
-  }
-}
