+++ kaldb/src/main/java/com/slack/kaldb/logstore/search/SearchResult.java
@@ -1,7 +1,6 @@
 package com.slack.kaldb.logstore.search;
 
 import com.google.common.base.Objects;
-import com.slack.kaldb.histogram.HistogramBucket;
 import com.slack.kaldb.logstore.LogMessage;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -11,7 +10,7 @@ import org.opensearch.search.aggregations.InternalAggregation;
 public class SearchResult<T> {
 
   private static final SearchResult EMPTY =
-      new SearchResult<>(Collections.emptyList(), 0, 0, Collections.emptyList(), 1, 1, 0, 0, null);
+      new SearchResult<>(Collections.emptyList(), 0, 0, 1, 1, 0, 0, null);
 
   public final long totalCount;
 
@@ -19,10 +18,6 @@ public class SearchResult<T> {
   // An iterator helps with the early termination of a search and may be efficient in some cases.
   public final List<T> hits;
   public final long tookMicros;
-
-  // TODO: Make this list immutable?
-  // TODO: Instead of histogram bucket, return tuple.
-  public final List<HistogramBucket> buckets;
 
   public final int failedNodes;
   public final int totalNodes;
@@ -35,7 +30,6 @@ public class SearchResult<T> {
     this.hits = new ArrayList<>();
     this.tookMicros = 0;
     this.totalCount = 0;
-    this.buckets = new ArrayList<>();
     this.failedNodes = 0;
     this.totalNodes = 0;
     this.totalSnapshots = 0;
@@ -48,7 +42,6 @@ public class SearchResult<T> {
       List<T> hits,
       long tookMicros,
       long totalCount,
-      List<HistogramBucket> buckets,
       int failedNodes,
       int totalNodes,
       int totalSnapshots,
@@ -57,7 +50,27 @@ public class SearchResult<T> {
     this.hits = hits;
     this.tookMicros = tookMicros;
     this.totalCount = totalCount;
-    this.buckets = buckets;
+    this.failedNodes = failedNodes;
+    this.totalNodes = totalNodes;
+    this.totalSnapshots = totalSnapshots;
+    this.snapshotsWithReplicas = snapshotsWithReplicas;
+    this.internalAggregation = internalAggregation;
+  }
+
+  @Deprecated
+  public SearchResult(
+      List<T> hits,
+      long tookMicros,
+      long totalCount,
+      List<Object> buckets,
+      int failedNodes,
+      int totalNodes,
+      int totalSnapshots,
+      int snapshotsWithReplicas,
+      InternalAggregation internalAggregation) {
+    this.hits = hits;
+    this.tookMicros = tookMicros;
+    this.totalCount = totalCount;
     this.failedNodes = failedNodes;
     this.totalNodes = totalNodes;
     this.totalSnapshots = totalSnapshots;
@@ -77,7 +90,6 @@ public class SearchResult<T> {
         && totalSnapshots == that.totalSnapshots
         && snapshotsWithReplicas == that.snapshotsWithReplicas
         && Objects.equal(hits, that.hits)
-        && Objects.equal(buckets, that.buckets)
         && Objects.equal(internalAggregation, that.internalAggregation);
   }
 
@@ -90,8 +102,6 @@ public class SearchResult<T> {
         + hits
         + ", tookMicros="
         + tookMicros
-        + ", buckets="
-        + buckets
         + ", failedNodes="
         + failedNodes
         + ", totalNodes="
@@ -111,7 +121,6 @@ public class SearchResult<T> {
         totalCount,
         hits,
         tookMicros,
-        buckets,
         failedNodes,
         totalNodes,
         totalSnapshots,
