--- kaldb/src/test/java/com/slack/kaldb/logstore/OpensearchShimTest.java
@@ -1,16 +1,0 @@
-package com.slack.kaldb.logstore;
-
-import java.time.Instant;
-import org.junit.Test;
-import org.opensearch.search.aggregations.Aggregator;
-
-public class OpensearchShimTest {
-
-  @Test
-  public void test() throws Exception {
-
-    Aggregator aggregator = OpensearchShim.test(20, 0, Instant.now().toEpochMilli());
-
-    System.out.println(aggregator.name());
-  }
-}
