+++ kaldb/src/test/java/com/slack/kaldb/logstore/search/StatsCollectorTest.java
@@ -11,7 +11,6 @@ import static com.slack.kaldb.testlib.MetricsUtil.getTimerCount;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import brave.Tracing;
-import com.slack.kaldb.histogram.HistogramBucket;
 import com.slack.kaldb.logstore.LogMessage;
 import com.slack.kaldb.testlib.TemporaryLogStoreAndSearcherRule;
 import java.io.IOException;
@@ -65,12 +64,12 @@ public class StatsCollectorTest {
 
     assertThat(allIndexItems.hits.size()).isEqualTo(0);
     assertThat(allIndexItems.totalCount).isEqualTo(5);
-    assertThat(allIndexItems.buckets.size()).isEqualTo(5);
-
-    for (HistogramBucket bucket : allIndexItems.buckets) {
-      // TODO: Test bucket start and end ranges.
-      assertThat(bucket.getCount()).isEqualTo(1);
-    }
+    //    assertThat(allIndexItems.buckets.size()).isEqualTo(5);
+    //
+    //    for (HistogramBucket bucket : allIndexItems.buckets) {
+    //      // TODO: Test bucket start and end ranges.
+    //      assertThat(bucket.getCount()).isEqualTo(1);
+    //    }
 
     assertThat(getCount(MESSAGES_RECEIVED_COUNTER, strictLogStore.metricsRegistry)).isEqualTo(5);
     assertThat(getCount(MESSAGES_FAILED_COUNTER, strictLogStore.metricsRegistry)).isEqualTo(0);
