+++ kaldb/src/test/java/com/slack/kaldb/server/SearchResultTest.java
@@ -3,16 +3,20 @@ package com.slack.kaldb.server;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import brave.Tracing;
-import com.slack.kaldb.histogram.HistogramBucket;
 import com.slack.kaldb.logstore.LogMessage;
+import com.slack.kaldb.logstore.opensearch.OpensearchShim;
 import com.slack.kaldb.logstore.search.SearchResult;
 import com.slack.kaldb.logstore.search.SearchResultUtils;
 import com.slack.kaldb.proto.service.KaldbSearch;
 import com.slack.kaldb.testlib.MessageUtil;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 import java.util.Random;
 import org.junit.Test;
+import org.opensearch.search.DocValueFormat;
+import org.opensearch.search.aggregations.InternalAggregation;
+import org.opensearch.search.aggregations.metrics.InternalAvg;
 
 public class SearchResultTest {
 
@@ -31,11 +35,11 @@ public class SearchResultTest {
       logMessages.add(logMessage);
     }
 
-    List<HistogramBucket> buckets = new ArrayList<>();
-    buckets.add(new HistogramBucket(1, 2));
+    InternalAggregation internalAggregation =
+        new InternalAvg("avg", 10, 10, DocValueFormat.RAW, Map.of("foo", "bar"));
 
     SearchResult<LogMessage> searchResult =
-        new SearchResult<>(logMessages, 1, 1000, buckets, 1, 5, 7, 7, null);
+        new SearchResult<>(logMessages, 1, 1000, 1, 5, 7, 7, internalAggregation);
     KaldbSearch.SearchResult protoSearchResult =
         SearchResultUtils.toSearchResultProto(searchResult);
 
@@ -46,7 +50,8 @@ public class SearchResultTest {
     assertThat(protoSearchResult.getTotalNodes()).isEqualTo(5);
     assertThat(protoSearchResult.getTotalSnapshots()).isEqualTo(7);
     assertThat(protoSearchResult.getSnapshotsWithReplicas()).isEqualTo(7);
-    assertThat(protoSearchResult.getBucketsCount()).isEqualTo(1);
+    assertThat(protoSearchResult.getInternalAggregations().toByteArray())
+        .isEqualTo(OpensearchShim.toByteArray(internalAggregation));
 
     SearchResult<LogMessage> convertedSearchResult =
         SearchResultUtils.fromSearchResultProto(protoSearchResult);
